+++ kotlin-stdlib/common/main/kotlin/com/javiersc/kotlin/stdlib/validate/Validator.kt
@@ -2,54 +2,33 @@ package com.javiersc.kotlin.stdlib.validate
 
 import com.javiersc.kotlin.stdlib.Either
 import com.javiersc.kotlin.stdlib.UseContextParameters
+import com.javiersc.kotlin.stdlib.left
+import com.javiersc.kotlin.stdlib.right
 
-public class Validator<V : Any>
+public class Validator<E, V>
 @PublishedApi
 internal constructor(
     @PublishedApi internal val name: String,
-    @PublishedApi internal val block: ValidatorScope<V>.(V) -> Unit,
+    @PublishedApi internal val block: ValidatorScope<E, V>.(V) -> Unit,
+    private val _unit0: Unit,
 ) {
 
-    public inline fun <reified A : V> validate(value: A): Either<String, A> {
-        val validatorScope = ValidatorScope<V>()
+    public fun <A : V> validate(value: A): Either<List<E>, A> {
+        val validatorScope = ValidatorScope<E, V>()
         block(validatorScope, value)
 
-        val otherwises: List<String> =
+        val otherwises: List<E> =
             validatorScope.validatables.mapNotNull { if (it.predicate()) it.otherwise() else null }
 
-        if (otherwises.isEmpty()) return Either.Right(value)
+        if (otherwises.isEmpty()) return value.right()
 
-        val sanitizedOtherwises: List<String> = buildList {
-            addAll(otherwises)
-
-            if (otherwises.count { it.contains(InvalidProperty) } > 1) {
-                removeAll { it.contains(InvalidProperty) }
-                add(MultipleInvalidProperties)
-            }
-
-            if (otherwises.count { it.contains(EmptyProperty) } > 1) {
-                removeAll { it.contains(EmptyProperty) }
-                add(MultipleEmptyProperties)
-            }
-        }
-
-        val title = "'$name' is invalid due to:\n"
-        val otherwisesAsText: String =
-            sanitizedOtherwises.joinToString("\n") { message -> "    - $message" }
-        val errorMessage: String = title + otherwisesAsText
-        return Either.Left(value = errorMessage)
-    }
-
-    public companion object {
-
-        public inline operator fun <reified T : Any> invoke(
-            name: String = T::class.simpleName ?: "Validator",
-            @UseContextParameters noinline block: ValidatorScope<T>.(T) -> Unit,
-        ): Lazy<Validator<T>> = lazy { Validator(name = name, block = block) }
+        val either: Either<List<E>, A> = otherwises.left()
+        return either
     }
 }
 
-@PublishedApi internal const val InvalidProperty: String = "Invalid property"
-@PublishedApi internal const val MultipleInvalidProperties: String = "Multiple invalid properties"
-@PublishedApi internal const val EmptyProperty: String = "Empty property"
-@PublishedApi internal const val MultipleEmptyProperties: String = "Multiple empty properties"
+@Suppress("FunctionName")
+public inline fun <E : Any, reified T : Any> Validator(
+    name: String = T::class.simpleName ?: "Validator",
+    @UseContextParameters noinline block: ValidatorScope<E, T>.(T) -> Unit,
+): Lazy<Validator<E, T>> = lazy { Validator(name = name, block = block, _unit0 = Unit) }