+++ kotlin-stdlib/common/main/kotlin/com/javiersc/kotlin/stdlib/validate/validate.kt
@@ -2,7 +2,9 @@ package com.javiersc.kotlin.stdlib.validate
 
 import com.javiersc.kotlin.stdlib.Either
 
-public inline fun <reified T : Any> T.validateWith(validator: Validator<T>): Either<String, T> {
-    val validated: Either<String, T> = validator.validate(this)
+public inline fun <reified E : Any, reified T : Any> T.validateWith(
+    validator: Validator<E, T>
+): Either<List<E>, T> {
+    val validated: Either<List<E>, T> = validator.validate(this)
     return validated
 }