+++ kotlin-stdlib/common/test/kotlin/com/javiersc/kotlin/stdlib/validate/ValidatorValidIfTest.kt
@@ -11,34 +11,48 @@ class ValidatorValidIfTest {
 
     @Test
     fun given__a_validator_using_validIf_api_and_a_book__when__validates__then__is_valid() {
-        val validator: Validator<Book> by Validator { book ->
-            validIf { book.title == "Lord of the Rings" }
-            validIf { book.author.name == "Tolkien" }
+        val validator: Validator<Book.Error, Book> by Validator { book ->
+            validIf(
+                predicate = { book.title == "Lord of the Rings" },
+                otherwise = { Book.Error.WrongTitle },
+            )
+            validIf(
+                predicate = { book.author.name == "Tolkien" },
+                otherwise = { Book.Error.WrongAuthor },
+            )
         }
         LordOfTheRingsBook.validateWith(validator).assertRight()
     }
 
     @Test
     fun given__a_validator_using_validIf_api_and_a_book__when__validates__then__is_invalid() {
-        val validator: Validator<Book> by Validator { book ->
-            validIf { book.title == "Lord of the Rings" }
-            validIf { book.author.name == "Tolkien" }
+        val validator: Validator<Book.Error, Book> by Validator { book ->
+            validIf(
+                predicate = { book.title == "Lord of the Rings" },
+                otherwise = { Book.Error.WrongTitle },
+            )
+            validIf(
+                predicate = { book.author.name == "Tolkien" },
+                otherwise = { Book.Error.WrongAuthor },
+            )
         }
-        HarryPotterBook.validateWith(validator).assertLeft {
-            """
-                |'Book' is invalid due to:
-                |    - Multiple invalid properties
-            """
-        }
+        HarryPotterBook.validateWith(validator)
+            .assertLeft(Book.Error.WrongTitle, Book.Error.WrongAuthor)
     }
 
     @Test
     fun given__a_validator_using_validIf_api_with_invoke_lambda_and_a_book__when__validates__then__is_valid() {
-        val validator: Validator<Book> by Validator { book ->
+        val validator: Validator<Book.Error, Book> by Validator { book ->
             book {
-                validIf { title == "Lord of the Rings" }
-                author { //
-                    validIf { name == "Tolkien" }
+                validIf(
+                    predicate = { title == "Lord of the Rings" },
+                    otherwise = { Book.Error.WrongTitle },
+                )
+                author {
+                    validIf(
+                        predicate = { name == "Tolkien" },
+                        otherwise = { Book.Error.WrongAuthor },
+                    )
                 }
             }
         }
@@ -47,19 +61,21 @@ class ValidatorValidIfTest {
 
     @Test
     fun given__a_validator_using_validIf_api_with_invoke_lambda_and_a_book__when__validates__then__is_invalid() {
-        val validator: Validator<Book> by Validator { book ->
+        val validator: Validator<Book.Error, Book> by Validator { book ->
             book {
-                validIf { title == "Lord of the Rings" }
-                author { //
-                    validIf { name == "Tolkien" }
+                validIf(
+                    predicate = { title == "Lord of the Rings" },
+                    otherwise = { Book.Error.WrongTitle },
+                )
+                author {
+                    validIf(
+                        predicate = { name == "Tolkien" },
+                        otherwise = { Book.Error.WrongAuthor },
+                    )
                 }
             }
         }
-        HarryPotterBook.validateWith(validator).assertLeft {
-            """
-                |'Book' is invalid due to:
-                |    - Multiple invalid properties
-            """
-        }
+        HarryPotterBook.validateWith(validator)
+            .assertLeft(Book.Error.WrongTitle, Book.Error.WrongAuthor)
     }
 }