+++ kotlin-stdlib/common/test/kotlin/com/javiersc/kotlin/stdlib/validate/ValidatorValidTest.kt
@@ -11,8 +11,8 @@ class ValidatorValidTest {
 
     @Test
     fun given__a_validator_using_valid_api_and_a_book__when__validates__then__is_valid() {
-        val validator: Validator<Book> by
-            Validator.invoke(name = "Book") { book ->
+        val validator: Validator<String, Book> by
+            Validator(name = "Book") { book ->
                 valid(
                     predicate = { book.title == "Lord of the Rings" },
                     otherwise = { "Title must be Lord of the Rings" },
@@ -27,38 +27,32 @@ class ValidatorValidTest {
 
     @Test
     fun given__a_validator_using_valid_api_and_a_book__when__validates__then__is_invalid() {
-        val validator: Validator<Book> by Validator { book ->
-            valid(
-                predicate = { book.title == "Lord of the Rings" },
-                otherwise = { "Title must be Lord of the Rings" },
-            )
-            valid(
-                predicate = { book.author.name == "Tolkien" },
-                otherwise = { "Author must be Tolkien" },
-            )
-        }
-        HarryPotterBook.validateWith(validator).assertLeft {
-            """
-                |'Book' is invalid due to:
-                |    - Title must be Lord of the Rings
-                |    - Author must be Tolkien
-            """
-        }
+        val validator: Validator<Book.Error, Book> by
+            Validator(name = "Book") { book ->
+                valid(
+                    predicate = { book.title == "Lord of the Rings" },
+                    otherwise = { Book.Error.WrongTitle },
+                )
+                valid(
+                    predicate = { book.author.name == "Tolkien" },
+                    otherwise = { Book.Error.WrongAuthor },
+                )
+            }
+
+        HarryPotterBook.validateWith(validator)
+            .assertLeft(Book.Error.WrongTitle, Book.Error.WrongAuthor)
     }
 
     @Test
     fun given__a_validator_using_valid_api_with_invoke_lambda_and_a_book__when__validates__then__is_valid() {
-        val validator: Validator<Book> by Validator { book ->
+        val validator: Validator<Book.Error, Book> by Validator { book ->
             book {
                 valid(
                     predicate = { title == "Lord of the Rings" },
-                    otherwise = { "Title must be Lord of the Rings" },
+                    otherwise = { Book.Error.WrongTitle },
                 )
                 author {
-                    valid(
-                        predicate = { name == "Tolkien" },
-                        otherwise = { "Author must be Tolkien" },
-                    )
+                    valid(predicate = { name == "Tolkien" }, otherwise = { Book.Error.WrongAuthor })
                 }
             }
         }
@@ -67,26 +61,18 @@ class ValidatorValidTest {
 
     @Test
     fun given__a_validator_using_valid_api_with_invoke_lambda_and_a_book__when__validates__then__is_invalid() {
-        val validator: Validator<Book> by Validator { book ->
+        val validator: Validator<Book.Error, Book> by Validator { book ->
             book {
                 valid(
                     predicate = { title == "Lord of the Rings" },
-                    otherwise = { "Title must be Lord of the Rings" },
+                    otherwise = { Book.Error.WrongTitle },
                 )
                 author {
-                    valid(
-                        predicate = { name == "Tolkien" },
-                        otherwise = { "Author must be Tolkien" },
-                    )
+                    valid(predicate = { name == "Tolkien" }, otherwise = { Book.Error.WrongAuthor })
                 }
             }
         }
-        HarryPotterBook.validateWith(validator).assertLeft {
-            """
-                |'Book' is invalid due to:
-                |    - Title must be Lord of the Rings
-                |    - Author must be Tolkien
-            """
-        }
+        HarryPotterBook.validateWith(validator)
+            .assertLeft(Book.Error.WrongTitle, Book.Error.WrongAuthor)
     }
 }